Add studentNumber, studentApproved and teacherId parameters to synchronization log. Move log service to correct level in project hierarchy. Add realisation id:s to import log messages
