+++ src/main/java/fi/helsinki/moodi/service/importing/ImportingService.java
@@ -28,7 +28,7 @@ import fi.helsinki.moodi.service.converter.CourseConverter;
 import fi.helsinki.moodi.service.course.Course;
 import fi.helsinki.moodi.service.course.CourseService;
 import fi.helsinki.moodi.service.dto.CourseDto;
-import fi.helsinki.moodi.service.synchronize.log.LoggingService;
+import fi.helsinki.moodi.service.log.LoggingService;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
